Made requested UI changes. Removed edit user options modal and replaced with simple folder icon button. Behavior questions remain on re-binding document list after change or restart app.
