+++ src/RoslynPad/App.config
@@ -41,9 +41,6 @@
       <setting name="DocumentPath" serializeAs="String">
         <value />
       </setting>
-      <setting name="CreateSamples" serializeAs="String">
-        <value>True</value>
-      </setting>
     </RoslynPad.Properties.Settings>
   </userSettings>
 </configuration>
