+++ src/RoslynPad/Properties/Settings.settings
@@ -23,8 +23,5 @@
     <Setting Name="DocumentPath" Type="System.String" Scope="User">
       <Value Profile="(Default)" />
     </Setting>
-    <Setting Name="CreateSamples" Type="System.Boolean" Scope="User">
-      <Value Profile="(Default)">True</Value>
-    </Setting>
   </Settings>
 </SettingsFile>
