--- src/RoslynPad/UserSettingsDialog.xaml
@@ -1,56 +1,0 @@
-﻿<UserControl x:Class="RoslynPad.UserSettingsDialog"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-             xmlns:local="clr-namespace:RoslynPad"
-             mc:Ignorable="d"
-             d:DataContext="{d:DesignInstance local:UserSettingsDialog}"
-             MinWidth="300"
-             BorderThickness="1"
-             BorderBrush="Gray"
-             Background="White">
-    <StackPanel Margin="10">
-        <TextBlock Text="User Settings"
-                   Style="{StaticResource Title}" />
-
-        <TextBlock Text="Document Path:"
-                       Margin="0,10,0,5" />
-
-        <StackPanel Orientation="Horizontal">
-            <TextBox Name="DocumentPath"
-                     Width="400"
-                     IsEnabled="False"
-                      />
-
-            <Button IsDefault="True"
-                    Style="{StaticResource DialogButton}"
-                    Name="DocumentPathBrowse"
-                    
-                    Content="_Browse"
-                    Click="DocumentPathBrowse_Click" />
-        </StackPanel>
-
-        <StackPanel Orientation="Horizontal" VerticalAlignment="Top" Margin="0, 20, 0, 0">
-            <CheckBox Name="CreateSampleFiles" 
-                      Margin="0, 0, 10, 0" 
-                      Click="CreateSampleFiles_Click" />
-            <TextBlock Text="Create Sample Files"
-                       Margin="0,0,0,0" />
-        </StackPanel>
-
-        <StackPanel HorizontalAlignment="Right"
-                    Orientation="Horizontal"
-                    Margin="50,30,0,0">
-            <Button IsDefault="True"
-                    Style="{StaticResource DialogButton}"
-                    Name="SaveButton"
-                    Content="_Save &amp; Restart"
-                    Click="Save_Click" />
-            <Button Content="_Cancel"
-                    Style="{StaticResource DialogButton}"
-                    IsCancel="True"
-                    Click="Cancel_Click" />
-        </StackPanel>
-    </StackPanel>
-</UserControl>
